Split qp into files
